DefaultOMEROService: check for registered objects
ImageJ tracks certain kinds of objects using the SciJava object service.
This includes instances of all SingletonPlugin subtypes, such as
CalculatorOp and ThresholdMethod.

The ImageJ UI will present the a dropdown box of available singletons of
that plugin type, from which the user can choose. For example,
ImageCalculator uses a CalculatorOp parameter to easily provide the user
with a selection of all available binary operations.

If the OMERO parameter value is a string, we can check the object
service for available objects whose string representation matches the
string given from OMERO. This allows ImageJ-OMERO to handle parameters
of singleton plugin types, among other situations.